Send data when settings frame is processed
While it is normally the case that data can be sent even before
receiving the server's settings frame, we do not want that in this case. We
want to first process the settings frame so that we can send as much
data as possible if servers want to allow more than 100 concurrent
streams (the default setting). See RFC 7540: "To avoid unnecessary
latency, clients are permitted to send additional frames to the server
immediately after sending the client connection preface, without
waiting to receive the server connection preface."