3rd shared libraries #2 (#77)
* Use json library from system if found

The devel package for library is packaged both for Debian
(nlohmann-json-dev) and Fedora (json-devel). Update included path to
find both by default if present on builded system.

* Detect installed cpp-httplib and prefer system lib

Detect also cpp-httplib if usable. Fallback to bundled source if not
found. Hide details of detection, prevent warning when not available.
Not packaged on Debian or Fedora (yet).