Fix copyright notices in the code
Based on the commit history, the project was run by NSONE up to the
year 2022. This PR adds copyright notices to the files that have been
created up to that year.

There is some interest to continue project development outside of NSONE.
All new changes are attributed to Flamethrower Contributors.